Manage position state manually
This addresses three issues:

1. The device always sends a position state of STOPPED, even when the device is currently in motion
2. The initial behaviour was to sync the target and current positions whenever the device state gets updated
3. The target position was being improperly inverted between homebridge and the device

While (1) wasn't an especially big deal, (2) and (3) were. (2) was a problem because homekit sends multiple requests to set the target position while dragging the position UI, and each one causes the device to stop and restart its movement - and each time it stopped, it would send a Report message to broadcast its new position. This would be received which would then sync the target position to the current position, causing homekit to think that the movement had reached its final position. (3) was a problem because homekit would also then think that the target position had changed and that the device is in perpetual motion.

Managing the position state manually means the accessory can accurately detect when a movement has completed, even if its final position is not the same as the target position (which can happen due to small position differences, or when the movement has been interrupted externally, e.g. via the remote).